+++ Services/RecentServices.cs
@@ -24,7 +24,7 @@ public class RecentServices
         return songs;
     } 
     
-    public async Task AddRecent(string username, long id)
+    public async Task AddRecent(string username, int id)
     {
         await _database.Call($"INSERT INTO recents (username, song, added_date) " +
                              $"VALUES ('{username}', {id}, {DateTime.Now:yyyy-MM-dd H:mm:ss}) " +
